Adds The Frontend Cryptocurrency and Deploying to Production and Backend-Included Improvements.
Created toggle-able transaction displays.
Built a reusable frontend transaction component.
Added routing for a multi-page application that renders different content based on the url.
Supported the conducting of transactions through a form.
Posted transactions through the fetch API.
Visualized the transaction pool - and added real-time polling capabilities.
Added a way to mine a block through the frontend.
Deployed the project to production.
Added a known addresses feature with both the backend and frontend.
Supported a paginated exploration of the blocks with both the backend and frontend.